Allow status LED to indicate client connection
When the FOB is first powered on and no clients have connected, it will
start blinking the yellow status LED. when first client to connects and
announces that it is ready to receive commands, the FOB will note the
IP address of the client so that all future commands only go there. Also
the blinking yellow status LED is switched to a solid green.

The button LEDs will also indicate what scenario is currently running.
When the 'a' version is running, the LED will blink and while the 'b'
version is running the LED will remain solid.